refactor(web): use a real testimonial component
