+++ packages/web/src/lib/components/TestimonialCollection.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-import { Link } from 'components';
 import Testimonial from './Testimonial.svelte';
 
 type TestimonialItem = {
@@ -15,19 +14,42 @@ export let testimonials: TestimonialItem[] = [];
 const { class: extraClass = '', ...restProps } = $$restProps;
 </script>
 
-<section
-	{...restProps}
-	class={`mx-auto w-full max-w-6xl ${extraClass}`.trim()}
->
-	<div class="columns-1 gap-6 sm:columns-2 lg:columns-3">
+<section {...restProps} class={extraClass}>
+	<div class="testimonial-grid">
 		{#each testimonials as item, index (item.authorName + index)}
-			<Link href={item.source} class={`block mb-6 break-inside-avoid ${index % 2 == 0 ? "skew-hover" : "skew-hover-left"}`}>
+			<a class="testimonial-link" href={item.source}>
 				<Testimonial
 					authorName={item.authorName}
 					authorSubtitle={item.authorSubtitle}
 					testimonial={item.testimonial}
+					accent={index === 4}
 				/>
-			</Link>
+			</a>
 		{/each}
 	</div>
 </section>
+
+<style>
+.testimonial-grid {
+	display: grid;
+	grid-template-columns: repeat(3, minmax(0, 1fr));
+	gap: 1.1rem;
+}
+
+.testimonial-link {
+	color: inherit;
+	text-decoration: none;
+}
+
+.testimonial-link:hover :global(figure) {
+	border-color: var(--marketing-amber);
+	box-shadow: 0 10px 24px -16px rgba(28, 26, 22, 0.16);
+	transform: translateY(-1px);
+}
+
+@media (max-width: 880px) {
+	.testimonial-grid {
+		grid-template-columns: 1fr;
+	}
+}
+</style>
